4 Revert "MPT-10404 Termination Orders and Returns (#399)"
This reverts commit 092e0cd7bf89c762f73daa2380ba9f48e5c51ed6, reversing
changes made to b1c075a2fb50561dfa609e3af9caaffb81db6015.
