+++ adobe_vipm/adobe/constants.py
@@ -26,7 +26,6 @@ STATUS_LINE_ITEM_OFFER_ID_EXPIRED = "3123"
 STATUS_INTERNAL_SERVER_ERROR = "1124"
 STATUS_GC_DEPLOYMENT_ACTIVE = "1000"
 STATUS_SUBSCRIPTION_TERMINATED = "1004"
-STATUS_SUBSCRIPTION_ACTIVE = "1000"
 
 
 ORDER_STATUS_DESCRIPTION = {
