+++ adobe_vipm/flows/fulfillment/termination.py
@@ -11,7 +11,6 @@ from adobe_vipm.adobe.constants import STATUS_INVALID_RENEWAL_STATE
 from adobe_vipm.adobe.errors import AdobeAPIError
 from adobe_vipm.flows.constants import (
     ERR_INVALID_RENEWAL_STATE,
-    ERR_INVALID_TERMINATION_ORDER_QUANTITY,
     TEMPLATE_NAME_TERMINATION,
 )
 from adobe_vipm.flows.context import Context
@@ -30,7 +29,6 @@ from adobe_vipm.flows.pipeline import Pipeline, Step
 from adobe_vipm.flows.utils import (
     get_adobe_subscription_id,
     get_subscription_by_line_and_item_id,
-    validate_subscription_and_returnable_orders,
 )
 
 logger = logging.getLogger(__name__)
@@ -45,24 +43,15 @@ class GetReturnableOrders(Step):
         adobe_client = get_adobe_client()
         for line in context.downsize_lines:
             sku = line["item"]["externalIds"]["vendor"]
-            is_valid, returnable_orders = validate_subscription_and_returnable_orders(
-                adobe_client,
-                context,
-                line,
-                sku,
-                return_orders=context.adobe_return_orders.get(sku)
-            )
-            logger.info(f"{context}: returnable orders: {returnable_orders} for {sku}")
-            if not is_valid:
-                switch_order_to_failed(
-                    client,
-                    context.order,
-                    ERR_INVALID_TERMINATION_ORDER_QUANTITY.to_dict(),
+            context.adobe_returnable_orders[sku] = (
+                adobe_client.get_returnable_orders_by_sku(
+                    context.authorization_id,
+                    context.adobe_customer_id,
+                    sku,
+                    context.adobe_customer["cotermDate"],
+                    return_orders=context.adobe_return_orders.get(sku),
                 )
-                return
-
-            context.adobe_returnable_orders[sku] = returnable_orders
-
+            )
         returnable_orders_count = sum(
             len(v) for v in context.adobe_returnable_orders.values()
         )
@@ -140,6 +129,7 @@ def fulfill_termination_order(client, order):
         GetReturnableOrders(),
         ValidateDownsizes3YC(),
         SubmitReturnOrders(),
+        SwitchAutoRenewalOff(),
         CompleteOrder(TEMPLATE_NAME_TERMINATION),
     )
     context = Context(order=order)
