+++ adobe_vipm/flows/validation/base.py
@@ -2,11 +2,12 @@ import logging
 import traceback
 
 from adobe_vipm.adobe.client import get_adobe_client
-from adobe_vipm.flows import constants
 from adobe_vipm.flows.helpers import (
     populate_order_info,
 )
 from adobe_vipm.flows.utils import (
+    is_change_order,
+    is_purchase_order,
     is_transfer_order,
     is_transfer_validation_enabled,
     notify_unhandled_exception_in_teams,
@@ -19,7 +20,6 @@ from adobe_vipm.flows.validation.change import validate_change_order
 from adobe_vipm.flows.validation.purchase import (
     validate_purchase_order,
 )
-from adobe_vipm.flows.validation.termination import validate_termination_order
 from adobe_vipm.flows.validation.transfer import validate_transfer
 
 logger = logging.getLogger(__name__)
@@ -39,23 +39,18 @@ def validate_order(client, order):
     try:
         has_errors = False
 
-        match order["type"]:
-            case constants.ORDER_TYPE_PURCHASE:
-                if is_transfer_order(order) and is_transfer_validation_enabled(
-                    order
-                ):
-                    adobe_client = get_adobe_client()
-                    order = populate_order_info(client, order)
-                    order = reset_ordering_parameters_error(order)
-                    order = reset_order_error(order)
-                    has_errors, order = validate_transfer(client, adobe_client, order)
-                else:
-                    has_errors, order = validate_purchase_order(client, order)
-            case constants.ORDER_TYPE_CHANGE:
-                has_errors, order = validate_change_order(client, order)
-            case constants.ORDER_TYPE_TERMINATION:
-                has_errors, order = validate_termination_order(client, order)
-
+        if is_purchase_order(order):
+            has_errors, order = validate_purchase_order(client, order)
+        elif is_change_order(order):
+            has_errors, order = validate_change_order(client, order)
+        elif is_transfer_order(order) and is_transfer_validation_enabled(
+            order
+        ):  # pragma: no branch
+            adobe_client = get_adobe_client()
+            order = populate_order_info(client, order)
+            order = reset_ordering_parameters_error(order)
+            order = reset_order_error(order)
+            has_errors, order = validate_transfer(client, adobe_client, order)
 
         order = update_parameters_visibility(order)
