+++ tests/flows/fulfillment/test_transfer.py
@@ -157,7 +157,7 @@ def test_transfer(
         return_value=subscription,
     )
     mocked_get_onetime = mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=items_factory(item_id=2, external_vendor_id="99999999CA"),
     )
 
@@ -490,7 +490,7 @@ def test_transfer_with_no_profile_address(
         return_value=subscription,
     )
     mocked_get_onetime = mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=items_factory(item_id=2, external_vendor_id="99999999CA"),
     )
 
@@ -1306,7 +1306,7 @@ def test_fulfill_transfer_order_already_migrated(
     )
 
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=items_factory(item_id=2, external_vendor_id="99999999CA"),
     )
     product_items = items_factory()
@@ -1499,7 +1499,7 @@ def test_fulfill_transfer_order_with_no_profile_address_already_migrated(
     )
 
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=items_factory(item_id=2, external_vendor_id="99999999CA"),
     )
     product_items = items_factory()
@@ -1669,7 +1669,7 @@ def test_fulfill_transfer_order_already_migrated_error_order_line_updated(
     )
 
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=items_factory(item_id=2, external_vendor_id="99999999CA"),
     )
 
@@ -1820,7 +1820,7 @@ def test_fulfill_transfer_order_already_migrated_3yc(
     )
 
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=[],
     )
     product_items = items_factory()
@@ -2033,7 +2033,7 @@ def test_fulfill_transfer_order_already_migrated_(
     )
 
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=[],
     )
 
@@ -2318,7 +2318,7 @@ def test_transfer_3yc_customer(
         return_value=subscription,
     )
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=[],
     )
 
@@ -2613,7 +2613,7 @@ def test_transfer_3yc_customer_with_no_profile_address(
         return_value=subscription,
     )
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=[],
     )
 
@@ -2889,7 +2889,7 @@ def test_fulfill_transfer_order_already_migrated_all_items_expired_create_new_or
     )
 
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=items_factory(item_id=2, external_vendor_id="99999999CA"),
     )
 
@@ -3051,7 +3051,7 @@ def test_fulfill_transfer_order_already_migrated_empty_adobe_items(
     )
 
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=items_factory(item_id=2, external_vendor_id="99999999CA"),
     )
 
@@ -3280,7 +3280,7 @@ def test_transfer_gc_account_all_deployments_created(
         return_value=subscription,
     )
     mocked_get_onetime = mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=items_factory(item_id=2, external_vendor_id="99999999CA"),
     )
 
@@ -3627,7 +3627,7 @@ def test_transfer_gc_account_no_deployments(
         return_value=subscription,
     )
     mocked_get_onetime = mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=items_factory(item_id=2, external_vendor_id="99999999CA"),
     )
 
@@ -4467,7 +4467,7 @@ def test_transfer_gc_account_no_items_error_main_agreement(
         "adobe_vipm.flows.fulfillment.shared.update_agreement",
     )
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=items_factory(item_id=2, external_vendor_id="99999999CA"),
     )
 
@@ -4852,7 +4852,7 @@ def test_fulfill_transfer_gc_order_already_migrated_(
     )
 
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=[],
     )
 
@@ -5041,7 +5041,7 @@ def test_fulfill_transfer_gc_order_already_migrated_no_items_without_deployment(
     )
 
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=[],
     )
 
@@ -5240,7 +5240,7 @@ def test_transfer_gc_account_items_with_deployment_main_agreement(
     )
 
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=items_factory(item_id=2, external_vendor_id="99999999CA"),
     )
 
@@ -5392,7 +5392,7 @@ def test_transfer_gc_account_items_with_deployment_main_agreement_bulk_migrated(
     )
 
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=items_factory(item_id=2, external_vendor_id="99999999CA"),
     )
 
@@ -5604,7 +5604,7 @@ def test_transfer_gc_account_no_deployments_gc_parameters_updated(
         return_value=subscription,
     )
     mocked_get_onetime = mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=items_factory(item_id=2, external_vendor_id="99999999CA"),
     )
 
@@ -5970,7 +5970,7 @@ def test_transfer_gc_account_items_with_and_without_deployment_main_agreement_bu
     )
 
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=items_factory(item_id=2, external_vendor_id="99999999CA"),
     )
 
@@ -6089,7 +6089,7 @@ def test_fulfill_transfer_migrated_order_all_items_expired_add_new_item(
         return_value=product_items,
     )
     mocker.patch(
-        "adobe_vipm.flows.utils.order.get_product_onetime_items_by_ids",
+        "adobe_vipm.flows.utils.get_product_onetime_items_by_ids",
         return_value=items_factory(item_id=2, external_vendor_id="99999999CA"),
     )
